review comments fixing

+++ server/src/main/java/ru/practicum/admin/category/CategoryServiceImpl.java
@@ -4,30 +4,39 @@ import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import ru.practicum.exceptions.CategoryNotFoundException;
+import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.event.EventRepository;
+import ru.practicum.exceptions.OperationNotAllowedException;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
+@Transactional(readOnly = true)
 @Service
 @AllArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
+    @Transactional
     @Override
     public CategoryDto create(CategoryDto categoryDto) {
         return CategoryMapper.toDto(categoryRepository.save(CategoryMapper.toCategory(categoryDto)));
     }
 
+    @Transactional
     @Override
     public CategoryDto update(CategoryDto categoryDto) {
         return CategoryMapper.toDto(categoryRepository.save(CategoryMapper.toCategory(categoryDto)));
     }
 
+    @Transactional
     @Override
     public void remove(int catId) {
+        if (!eventRepository.findAllByCategoryId(catId).isEmpty()) {
+            throw new OperationNotAllowedException("Deleting a category with events is not allowed");
+        }
         categoryRepository.deleteById(catId);
     }
 
@@ -36,14 +45,5 @@ public class CategoryServiceImpl implements CategoryService {
         Page<Category> categoriesPage = categoryRepository.findAll(pageable);
         List<Category> categories = categoriesPage.getContent();
         return categories.stream().map(CategoryMapper::toDto).collect(Collectors.toList());
-    }
-
-    @Override
-    public Category getCategoryByIdOrThrowNotFound(int catId) {
-        Optional<Category> category = categoryRepository.findById(catId);
-        if (category.isEmpty()) {
-            throw new CategoryNotFoundException("Category not found");
-        }
-        return category.get();
     }
 }
